moving Tooltip variants to separate folder; separating FloatingComponent from main Tooltip component file
