+++ packages/react-components/src/components/Tooltip/components/UserGuide/SpotlightOverlay.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import cx from 'clsx';
-import styles from './Tooltip.module.scss';
+import styles from '../../Tooltip.module.scss';
 
 const baseClass = 'guide-tooltip';
