+++ packages/react-components/src/components/Tooltip/components/UserGuide/UserGuide.tsx
@@ -1,11 +1,12 @@
 import * as React from 'react';
-import { ModalPortalProps } from '../Modal/';
-import { ITooltipProps, Tooltip } from './Tooltip';
+import { ModalPortalProps } from '../../../Modal';
+import { Tooltip } from '../../Tooltip';
 import SpotlightOverlay from './SpotlightOverlay';
 import cx from 'clsx';
-import styles from './Tooltip.module.scss';
+import styles from '../../Tooltip.module.scss';
 import { ClientRectObject } from '@floating-ui/core';
 import VirtualReference from './virtualElementReference';
+import { ITooltipProps } from '../../types';
 
 const spotlightPadding = 8;
 const baseClass = 'guide-tooltip';
