+++ packages/react-components/src/components/Tooltip/components/UserGuide/index.ts
@@ -1,0 +1,1 @@
+export { UserGuide } from './UserGuide';
